Conflicts:
	oneflow/core/kernel/loss_kernel.h

+++ oneflow/core/kernel/normal_model_update_kernel.cuh
@@ -6,8 +6,8 @@ namespace oneflow {
 template<typename T>
 __host__ __device__ T RegularizeDiff(T diff, const T batch_instance_num, T l1, T l2,
                                      T pre_model_val) {
-  return diff / static_cast<int32_t>(batch_instance_num)
-         + l1 * ((pre_model_val >= 0) - (pre_model_val <= 0)) + l2 * pre_model_val;
+  return diff / batch_instance_num + l1 * ((pre_model_val >= 0) - (pre_model_val <= 0))
+         + l2 * pre_model_val;
 }
 
 }  // namespace oneflow
